+++ include/nmtools/utility.hpp
@@ -9,194 +9,12 @@
 #include <tuple>
 #include <type_traits>
 
+#include "nmtools/traits.hpp"
+
 namespace nmtools {
 
     template <int Num, int Denom>
     using ratio = std::ratio<Num,Denom>;
-
-    namespace traits {
-        template <typename D, typename = void>
-        struct is_std_array : std::false_type {};
-
-        template <typename D>
-        struct is_std_array<D, std::enable_if_t<
-            std::is_same_v<
-                D, std::array<typename D::value_type, std::tuple_size<D>::value>
-            > // is_same
-        > /* enable_if */ > : std::true_type {};
-
-        template <typename D, typename = void>
-        struct is_std_vector : std::false_type {};
-
-        template <typename D>
-        struct is_std_vector<D, std::enable_if_t<
-            std::is_same_v<
-                D, std::vector<typename D::value_type>
-            > // is_same
-        > /* enable_if */ > : std::true_type {};
-
-        template <typename D, typename = void>
-        struct has_push_back_op : std::false_type {};
-
-        template <typename D>
-        struct has_push_back_op<D, std::void_t<
-                decltype(std::declval<D>().push_back(std::declval<typename D::value_type>()))
-        > /* void_t */ > : std::true_type {};
-
-        template <typename T, typename = void>
-        struct is_std_complex : std::false_type {};
-
-        template <typename T>
-        struct is_std_complex<T, std::enable_if_t<
-            std::is_same_v< 
-                T, std::complex< typename T::value_type > 
-            > // is_same
-        > /* enable_if */ > : std::true_type {};
-
-        template <typename T, typename = void, typename = void, typename = void, typename = void>
-        struct is_insertable : std::false_type {};
-
-        template <typename T, typename U>
-        struct is_insertable<T, U, std::void_t< 
-            decltype(std::declval<T>().insert(std::declval<T>().begin(),std::declval<const U&>()))
-        >, void, void > : std::true_type {};
-
-        template <typename T, typename I, typename U>
-        struct is_insertable<T, I, U, std::void_t< 
-            decltype(std::declval<T>().insert(std::declval<I>(),std::declval<const U&>()))
-        >, void > : std::true_type {};
-
-        template <typename T, typename I, typename B, typename E>
-        struct is_insertable<T, I, B, E, std::void_t< 
-            decltype(std::declval<T>().insert(std::declval<I>(),std::declval<B>(),std::declval<E>()))
-        > > : std::true_type {};
-
-        template <typename T, typename = void>
-        struct is_clearable : std::false_type {};
-
-        template <typename T>
-        struct is_clearable<T, std::enable_if_t<
-            std::is_same_v<
-                decltype(std::declval<T>().clear()),void
-            > // is_same
-        > /* enable_if */ > : std::false_type {};
-
-        template <typename T, typename = void>
-        struct has_ref_square_bracket_operator : std::false_type {};
-
-        template <typename T>
-        struct has_ref_square_bracket_operator<T, std::enable_if_t< 
-            !std::is_const<decltype(std::declval<T>()[size_t{}])>::value &&
-            std::is_reference<decltype(std::declval<T>()[size_t{}])>::value
-        > > : std::true_type {};
-
-        template <typename T, typename = void>
-        struct is_2d_array : std::false_type {};
-
-        template <typename T>
-        struct is_2d_array<T, std::void_t<decltype(std::declval<T>()[0][0])> 
-        > : std::true_type {};
-
-        template <typename T, typename U, typename = void>
-        struct multiplicative : std::false_type {};
-
-        template <typename T, typename U>
-        struct multiplicative<T, U, std::void_t<
-            decltype(std::declval<T>() * std::declval<U>())
-        > > : std::true_type {};
-
-        template <typename T, typename U, typename = void>
-        struct additive : std::false_type {};
-
-        template <typename T, typename U>
-        struct additive<T, U, std::void_t<
-            decltype(std::declval<T>() + std::declval<U>())
-        > > : std::true_type {};
-
-        template <typename F, typename...Args>
-        struct is_callable {
-        private:
-            template <typename FN>
-            constexpr static auto test(int) -> decltype(std::declval<FN>()(std::declval<Args>()...), bool()) {
-                return true;
-            }
-            template <typename>
-            constexpr static auto test(...) -> decltype(bool()) {
-                return false;
-            }
-        public:
-            constexpr static bool value = test<F>(int{});
-        };
-
-        template <typename T, typename = void>
-        struct has_transpose_op : std::false_type {};
-
-        template <typename T>
-        struct has_transpose_op<T, std::void_t<
-            decltype(std::declval<T>().transpose())
-        > > : std::true_type {};
-
-        template <typename T, typename = void>
-        struct is_transposeable : std::false_type {};
-
-        template <typename T>
-        struct is_transposeable<T, std::enable_if_t< 
-            std::is_arithmetic_v<T> || has_transpose_op<T>::value 
-        > > : std::true_type {};
-
-        /* TODO : move (?) */
-        using std::begin;
-        using std::end;
-
-        template <typename T>
-        struct is_iterable {
-        private:
-            template <typename It>
-            static constexpr auto test(int)
-                -> decltype(begin(std::declval<It>()), end(std::declval<It>()), bool())
-            { return true; }
-            template <typename It>
-            static constexpr auto test(char)
-            { return false; }
-        public:
-            constexpr static bool value = test<T>(int());
-        };
-
-        template <typename T, typename = void>
-        struct is_indexable : std::false_type {};
-        template <typename T>
-        struct is_indexable<T, std::void_t<decltype(std::declval<T>()[size_t{}])> > : std::true_type {};
-
-        template <typename T, typename = void>
-        struct is_resizeable : std::false_type {};
-        template <typename T>
-        struct is_resizeable<T, 
-            std::void_t<decltype(std::declval<T>().resize(std::declval<typename T::size_type>()))> >
-        : std::true_type {};
-
-        template <typename T>
-        static constexpr auto is_resizeable_v = is_resizeable<T>::value;
-
-        template <typename ...T>
-        struct all_iterable {
-            constexpr static bool value = std::conjunction_v<is_iterable<T>...>;
-        };
-
-        template <typename ...T>
-        struct all_indexable {
-            constexpr static bool value = std::conjunction_v<is_indexable<T>...>;
-        };
-
-        template <typename ...T>
-        constexpr bool all_iterable_v = all_iterable<T...>::value;
-
-        template <typename ...T>
-        constexpr bool all_indexable_v = all_indexable<T...>::value;
-
-        template <typename T>
-        using is_std_array_or_vector = std::disjunction<is_std_array<std::decay_t<T>>,is_std_vector<std::decay_t<T>>>;
-
-    } // namespace traits
 
     /* TODO: remove */
     namespace mpl {
